requirements.txt
python-dotenv==1.2.2
click==8.3.1
rich==14.3.3
scrapy==2.14.1
Pillow==12.1.1
requests==2.32.5
streamlit==1.55.0
anthropic==0.84.0
torch
torchvision
transformers
opencv-python-headless
scikit-learn
scikit-image
numpy
scipy
plotly
